packages/workers/ingestion/pyproject.toml
[project]
name = "bibliotalk-ingestion"
version = "0.1.0"
description = "Bibliotalk Canon acquisition worker"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "httpx>=0.28",
    "openai>=1.60",
    "crawl4ai>=0.4",
    "yt-dlp>=2024.0",
    "feedparser>=6.0",
    "trafilatura>=2.0",
    "markitdown>=0.1",
    "python-multipart>=0.0.18",
    "python-slugify>=8.0",
    "sse-starlette>=2.0",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "httpx"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ingestion"]

[project.scripts]
ingestion = "ingestion.main:cli"
